Make refreshTickets refresh all fields data
The function refreshTickets currently finds all cells containing JIRA
keys, contacts Jira and replaces the cell contents with the same Jira
key and the issue status. But this is not what most users want. What
they want is to have all the filter data refreshed in the spreadsheet.

The function insertIssuesFromFilter(jsonFormData) is responsible for
inserting the results of a filter. Now it sets a note to the "summary"
cell containing a mark, a warning, and the original jsonFormData, from
which is can later know how to refresh the filter again.

The mark is the "[jira-tools]" string, allowing us to find later all
notes we inserted along the way. Then, when we want to refresh all
filters in a sheet we can use the getNotes method and look for the notes
containing the mark.

The warning is just a message telling the users to avoid editing or
removing the note itself, so that they will be able later on to refresh
the filter.

The jsonFormData is simply the JSON argumemt that was used to generate
the table.

Then, the refreshTickets function simply invoke insertIssuesFromFilter()
again with the same arguments and in the same cell to refresh the table.